requirements.txt
attrs==23.2.0
blinker==1.8.2
click==8.1.7
flask==3.0.3
gunicorn==22.0.0
importlib-metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.24.4
packaging==24.0
pluggy==0.13.1
py==1.11.0
pytest==6.2.4
python-dotenv==1.0.1
scikit-learn==1.3.2
scipy==1.10.1
threadpoolctl==3.5.0
toml==0.10.2
werkzeug==3.0.3
xgboost==2.0.3
zipp==3.19.2
